+++ packages/connectors/src/browser-scraper-utils.ts
@@ -83,6 +83,76 @@ export function validateCookieNotExpired(
 // -----------------------------------------------------------------------------
 
 /**
+ * Validates a URL is safe for server-side fetching.
+ * Blocks private/internal network addresses to prevent SSRF attacks.
+ *
+ * Returns silently when the URL is safe; throws with a descriptive message
+ * otherwise. Connectors that fetch URLs derived from remote/untrusted input
+ * (sitemaps, HN story links, RSS feeds configured by users, etc.) MUST call
+ * this at the trust boundary before issuing the request.
+ */
+export function validatePublicUrl(url: string): void {
+  let parsed: URL;
+  try {
+    parsed = new URL(url);
+  } catch {
+    throw new Error(`Invalid URL: ${url}`);
+  }
+
+  if (parsed.protocol !== 'https:' && parsed.protocol !== 'http:') {
+    throw new Error(`URL must use http: or https: protocol, got ${parsed.protocol}`);
+  }
+
+  const hostname = parsed.hostname.toLowerCase();
+
+  // Block localhost variants
+  if (hostname === 'localhost' || hostname === '[::1]' || hostname.endsWith('.localhost')) {
+    throw new Error(`URL must not point to localhost: ${hostname}`);
+  }
+
+  // IPv4 private/loopback/link-local/cloud-metadata ranges
+  const ipv4Match = hostname.match(/^(\d{1,3})\.(\d{1,3})\.(\d{1,3})\.(\d{1,3})$/);
+  if (ipv4Match) {
+    const [, a, b] = ipv4Match.map(Number);
+    if (
+      a === 127 || // loopback
+      a === 10 || // private
+      (a === 172 && b >= 16 && b <= 31) || // private
+      (a === 192 && b === 168) || // private
+      (a === 169 && b === 254) || // link-local incl. 169.254.169.254 cloud metadata
+      a === 0
+    ) {
+      throw new Error(`URL must not point to a private/internal IP address: ${hostname}`);
+    }
+  }
+
+  // IPv6 private ranges (bracketed notation)
+  if (hostname.startsWith('[')) {
+    const ipv6 = hostname.slice(1, -1).toLowerCase();
+    if (
+      ipv6 === '::1' ||
+      ipv6.startsWith('fe80:') || // link-local
+      ipv6.startsWith('fc') || // unique local
+      ipv6.startsWith('fd') ||
+      ipv6 === '::' ||
+      ipv6.startsWith('::ffff:') // IPv4-mapped IPv6
+    ) {
+      throw new Error(`URL must not point to a private/internal IPv6 address: ${hostname}`);
+    }
+  }
+
+  // Common internal hostnames
+  if (
+    hostname.endsWith('.internal') ||
+    hostname.endsWith('.local') ||
+    hostname.endsWith('.corp') ||
+    hostname.endsWith('.lan')
+  ) {
+    throw new Error(`URL must not point to an internal hostname: ${hostname}`);
+  }
+}
+
+/**
  * Validate that a URL is well-formed, uses HTTPS, and belongs to the expected
  * domain (hostname ends with `expectedDomain`).
  *
